+++ src/HLS/virtual_components/multiplier_conn_obj.hpp
@@ -54,7 +54,7 @@
 class multiplier_conn_obj : public generic_obj
 {
    /// number of bit of in/out ports
-   unsigned long long bitsize;
+   unsigned int bitsize;
 
    /// when true the multiplication is a multiplication to a constant
    bool multiply_by_constant_p;
@@ -79,7 +79,7 @@ class multiplier_conn_obj : public generic_obj
    /**
     * add a size to the component
     */
-   void add_bitsize(unsigned long long _bitsize)
+   void add_bitsize(unsigned int _bitsize)
    {
       bitsize = _bitsize > bitsize ? _bitsize : bitsize;
    }
@@ -87,7 +87,7 @@ class multiplier_conn_obj : public generic_obj
    /**
     * return the maximum bitsize associated with the component
     */
-   unsigned long long get_bitsize() const
+   unsigned int get_bitsize() const
    {
       return bitsize;
    }
